moved FillContext to ConsensusContext
